# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amiyo-Go API",
    "version": "0.2.0",
    "description": "Typed API contract for the Amiyo-Go mobile platform."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "firebaseBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Firebase ID token"
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri",
            "default": "https://api.amiyo-go.local/problems/internal-error"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "correlationId": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code"
        ]
      },
...
      "IdentitySession": {
        "type": "object",
        "properties": {
          "principal": {
            "type": "object",
            "properties": {
              "userId": {
                "type": "string",
                "format": "uuid"
              },
              "roles": {
                "type": "array",
                "items": {
                  "type": "string",
                  "enum": [
                    "CUSTOMER",
                    "VENDOR_OWNER",
                    "VENDOR_MANAGER",
                    "VENDOR_STAFF",
                    "SUPPORT_AGENT",
                    "FINANCE_ADMIN",
                    "OPERATIONS_ADMIN",
                    "SUPER_ADMIN"
                  ]
                }
              },
              "vendorIds": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "uuid"
                },
                "default": []
              }
            },
            "required": [
              "userId",
              "roles"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "DEACTIVATED",
              "ANONYMIZED"
            ]
          },
          "email": {
            "type": [
              "string",
              "null"
            ],
            "format": "email"
          },
          "phone": {
            "type": [
              "string",
              "null"
            ]
          },
          "profile": {
            "type": "object",
            "properties": {
              "firstName": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "lastName": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "displayName": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "avatarStorageKey": {
                "type": [
                  "string",
                  "null"
                ]
              },
              "locale": {
                "type": "string"
              },
              "currency": {
                "type": "string",
                "minLength": 3,
                "maxLength": 3
              }
            },
            "required": [
              "firstName",
              "lastName",
              "displayName",
              "avatarStorageKey",
              "locale",
              "currency"
            ]
          },
          "permissions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "vendorMemberships": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "vendorId": {
                  "type": "string",
                  "format": "uuid"
                },
                "role": {
                  "type": "string",
                  "enum": [
                    "CUSTOMER",
                    "VENDOR_OWNER",
                    "VENDOR_MANAGER",
                    "VENDOR_STAFF",
                    "SUPPORT_AGENT",
                    "FINANCE_ADMIN",
                    "OPERATIONS_ADMIN",
                    "SUPER_ADMIN"
                  ]
                },
                "permissions": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "vendorId",
                "role",
                "permissions"
              ]
            }
          }
        },
        "required": [
          "principal",
          "status",
          "email",
          "phone",
          "profile",
          "permissions",
          "vendorMemberships"
        ]
      },
      "Address": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "minLength": 1,
            "maxLength": 40
          },
          "recipientName": {
            "type": "string",
            "minLength": 2,
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "minLength": 8,
            "maxLength": 20
          },
          "line1": {
            "type": "string",
            "minLength": 3,
            "maxLength": 240
          },
          "line2": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 240
          },
          "division": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "district": {
            "type": "string",
            "minLength": 2,
            "maxLength": 80
          },
          "upazila": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 80
          },
          "unionName": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 80
          },
          "postalCode": {
            "type": [
              "string",
              "null"
            ],
            "maxLength": 20
          },
          "latitude": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": [
              "number",
              "null"
            ],
            "minimum": -180,
            "maximum": 180
          },
          "isDefault": {
            "type": "boolean",
            "default": false
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "label",
          "recipientName",
          "phone",
          "line1",
          "division",
          "district",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Device": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "installationId": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "appVersion": {
            "type": [
              "string",
              "null"
            ]
          },
          "revokedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "installationId",
          "platform",
          "appVersion",
          "revokedAt",
          "createdAt",
          "updatedAt"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/v2/auth/session": {
      "post": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Synchronized trusted user session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentitySession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/me": {
      "get": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile and authorization context",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentitySession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "lastName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "minLength": 1,
                    "maxLength": 80
                  },
                  "displayName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "en",
                      "bn"
                    ]
                  },
                  "currency": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentitySession"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/me/addresses": {
      "get": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user addresses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Address"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "label": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "recipientName": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 120
                  },
                  "phone": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 20
                  },
                  "line1": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 240
                  },
                  "line2": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 240
                  },
                  "division": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "district": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 80
                  },
                  "upazila": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 80
                  },
                  "unionName": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 80
                  },
                  "postalCode": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 20
                  },
                  "latitude": {
                    "type": [
                      "number",
                      "null"
                    ],
                    "minimum": -90,
                    "maximum": 90
                  },
                  "longitude": {
                    "type": [
                      "number",
                      "null"
                    ],
                    "minimum": -180,
                    "maximum": 180
                  },
                  "isDefault": {
                    "type": "boolean",
                    "default": false
                  }
                },
                "required": [
                  "label",
                  "recipientName",
                  "phone",
                  "line1",
                  "division",
                  "district"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created address",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Address"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/devices": {
      "get": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Registered devices",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "firebaseBearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "installationId": {
                    "type": "string",
                    "minLength": 8,
                    "maxLength": 200
                  },
                  "platform": {
                    "type": "string",
                    "enum": [
                      "android",
                      "ios",
                      "web"
                    ]
                  },
                  "appVersion": {
                    "type": [
                      "string",
                      "null"
                    ],
                    "maxLength": 40
                  },
                  "pushToken": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 512
                  },
                  "pushProvider": {
                    "type": "string",
                    "enum": [
                      "expo",
                      "fcm",
                      "apns"
                    ]
                  }
                },
                "required": [
                  "installationId",
                  "platform"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registered device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permission",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}